Centralisé les transformations des données dans le view model
